+++ src/test/java/org/opensearch/knn/index/codec/KNN990Codec/NativeEngines990KnnVectorsReaderTests.java
@@ -7,20 +7,25 @@ package org.opensearch.knn.index.codec.KNN990Codec;
 
 import com.google.common.collect.ImmutableSet;
 import lombok.SneakyThrows;
+import org.apache.lucene.codecs.hnsw.FlatVectorsReader;
+import org.apache.lucene.index.ByteVectorValues;
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.FieldInfos;
 import org.apache.lucene.index.SegmentInfo;
 import org.apache.lucene.index.SegmentReadState;
+import org.apache.lucene.index.VectorEncoding;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.store.IndexInput;
 import org.mockito.MockedStatic;
+import org.mockito.Mockito;
 import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.index.codec.KNNCodecTestUtil;
 import org.opensearch.knn.index.engine.KNNEngine;
 import org.opensearch.knn.memoryoptsearch.VectorSearcher;
 import org.opensearch.knn.memoryoptsearch.VectorSearcherFactory;
 
+import java.io.IOException;
 import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.Collections;
@@ -31,8 +36,10 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.opensearch.knn.common.KNNConstants.KNN_ENGINE;
+import static org.opensearch.knn.common.KNNConstants.QFRAMEWORK_CONFIG;
 import static org.opensearch.knn.index.mapper.KNNVectorFieldMapper.KNN_FIELD;
 
 public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
@@ -43,10 +50,74 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
         final FieldInfos fieldInfos = new FieldInfos(fieldInfoArray);
 
         // Load vector searchers
-        final NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, Collections.emptySet());
+        final NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, Collections.emptySet(), null);
 
         final Map<String, NativeEngines990KnnVectorsReader.VectorSearcherHolder> vectorSearchers = getVectorSearcherHolders(reader);
         assertTrue(vectorSearchers.isEmpty());
+    }
+
+    public void testFlatVectorReaderIsCalled_whenNoQuantization() throws IOException {
+        FieldInfo fieldInfo = KNNCodecTestUtil.FieldInfoBuilder.builder("field1")
+            .fieldNumber(0)
+            .addAttribute(KNN_ENGINE, KNNEngine.FAISS.getName())
+            .addAttribute(KNNVectorFieldMapper.KNN_FIELD, "true")
+            .vectorEncoding(VectorEncoding.FLOAT32)
+            .build();
+
+        FieldInfo[] fieldInfoArray = new FieldInfo[] { fieldInfo };
+        final FieldInfos fieldInfos = new FieldInfos(fieldInfoArray);
+
+        final FlatVectorsReader flatVectorsReader = mock(FlatVectorsReader.class);
+        NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, Set.of(), flatVectorsReader);
+        reader.getByteVectorValues("field1");
+        verify(flatVectorsReader).getByteVectorValues("field1");
+    }
+
+    public void testBinaryVectorValuesIsCalled_whenQuantizationIsAvailable_thenSuccess() throws IOException {
+        FieldInfo fieldInfo = KNNCodecTestUtil.FieldInfoBuilder.builder("field1")
+            .fieldNumber(0)
+            .addAttribute(KNN_ENGINE, KNNEngine.FAISS.getName())
+            .addAttribute(QFRAMEWORK_CONFIG, "some-value")
+            .addAttribute(KNNVectorFieldMapper.KNN_FIELD, "true")
+            .build();
+
+        FieldInfo[] fieldInfoArray = new FieldInfo[] { fieldInfo };
+        final FieldInfos fieldInfos = new FieldInfos(fieldInfoArray);
+
+        KNNEngine mockFaiss = spy(KNNEngine.FAISS);
+        VectorSearcherFactory mockFactory = mock(VectorSearcherFactory.class);
+        VectorSearcher mockSearcher = mock(VectorSearcher.class);
+        when(mockSearcher.getByteVectorValues()).thenReturn(mock(ByteVectorValues.class));
+        when(mockFaiss.getVectorSearcherFactory()).thenReturn(mockFactory);
+        when(mockFactory.createVectorSearcher(any(), any(), any(), any())).thenReturn(mockSearcher);
+
+        final FlatVectorsReader flatVectorsReader = mock(FlatVectorsReader.class);
+        try (MockedStatic<KNNEngine> mockedStatic = mockStatic(KNNEngine.class)) {
+            mockedStatic.when(() -> KNNEngine.getEngine(any())).thenReturn(mockFaiss);
+            final Set<String> filesInSegment = Set.of("_0_165_field1.faiss");
+            mockedStatic.when(KNNEngine::getEnginesThatCreateCustomSegmentFiles).thenReturn(ImmutableSet.of(mockFaiss));
+            NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, filesInSegment, flatVectorsReader);
+            reader.getByteVectorValues("field1");
+            verify(mockSearcher).getByteVectorValues();
+            verify(flatVectorsReader, Mockito.never()).getByteVectorValues("field1");
+        }
+    }
+
+    public void testFlatVectorReaderIsCalled_whenEncodingIsByte_thenSuccess() throws IOException {
+        FieldInfo fieldInfo = KNNCodecTestUtil.FieldInfoBuilder.builder("field1")
+            .fieldNumber(0)
+            .addAttribute(KNN_ENGINE, KNNEngine.FAISS.getName())
+            .addAttribute(KNNVectorFieldMapper.KNN_FIELD, "true")
+            .vectorEncoding(VectorEncoding.BYTE)
+            .build();
+
+        FieldInfo[] fieldInfoArray = new FieldInfo[] { fieldInfo };
+        final FieldInfos fieldInfos = new FieldInfos(fieldInfoArray);
+
+        final FlatVectorsReader flatVectorsReader = mock(FlatVectorsReader.class);
+        NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, Set.of(), flatVectorsReader);
+        reader.getByteVectorValues("field1");
+        verify(flatVectorsReader).getByteVectorValues("field1");
     }
 
     @SneakyThrows
@@ -90,7 +161,7 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
             mockedStatic.when(KNNEngine::getEnginesThatCreateCustomSegmentFiles).thenReturn(ImmutableSet.of(mockFaiss));
 
             // Load reader
-            final NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, filesInSegment);
+            final NativeEngines990KnnVectorsReader reader = createReader(fieldInfos, filesInSegment, null);
 
             // Check the size of `vectorSearcher`
             final Map<String, NativeEngines990KnnVectorsReader.VectorSearcherHolder> vectorSearchers = getVectorSearcherHolders(reader);
@@ -154,6 +225,17 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
 
     @SneakyThrows
     private static NativeEngines990KnnVectorsReader createReader(final FieldInfos fieldInfos, final Set<String> filesInSegment) {
+        return createReader(fieldInfos, filesInSegment, null);
+    }
+
+    @SneakyThrows
+    private static NativeEngines990KnnVectorsReader createReader(final FieldInfos fieldInfos, final Set<String> filesInSegment) {
+        return createReader(fieldInfos, filesInSegment, null);
+    }
+        final FieldInfos fieldInfos,
+        final Set<String> filesInSegment,
+        final FlatVectorsReader flatVectorsReader
+    ) {
         // Prepare infra
         final IndexInput mockIndexInput = mock(IndexInput.class);
         final Directory mockDirectory = mock(Directory.class);
@@ -164,7 +246,7 @@ public class NativeEngines990KnnVectorsReaderTests extends KNNTestCase {
         final SegmentReadState readState = new SegmentReadState(mockDirectory, segmentInfo, fieldInfos, IOContext.DEFAULT);
 
         // Create reader
-        final NativeEngines990KnnVectorsReader reader = new NativeEngines990KnnVectorsReader(readState, null);
+        final NativeEngines990KnnVectorsReader reader = new NativeEngines990KnnVectorsReader(readState, flatVectorsReader);
         return reader;
     }